added (very broken) if/else, and/or statements, and semi-implemented while loops

+++ src/include/vm.h
@@ -6,7 +6,7 @@
 
 typedef struct {
     Chunk* chunk;
-    uint8_t* ip;
+    uint32_t* ip;
     Value* stack;
     int stackCount;
     int stackCapacity;
